feature(authentication): rename JS file
